pip: Move package distribution logic into a separate module
Move all related functionality related to package formats/distributions
(sdist, wheel) into a new module package_distributions.py

Also, move the corresponding tests into a new module.
